Added quick stats button functionality
