+++ app/build.gradle
@@ -30,7 +30,9 @@ dependencies {
     compile 'com.android.support:support-v4:25.2.0'
     compile 'com.google.firebase:firebase-database:10.2.1'
     compile 'com.google.firebase:firebase-auth:10.2.1'
+    compile 'com.android.support.constraint:constraint-layout:1.0.0-beta2'
     testCompile 'junit:junit:4.12'
 }
+
 
 apply plugin: 'com.google.gms.google-services'
