+++ app/src/main/AndroidManifest.xml
@@ -12,9 +12,9 @@
 
     <application
         android:allowBackup="true"
-        android:largeHeap="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:largeHeap="true"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
@@ -36,11 +36,11 @@
         <activity android:name=".activities.WorkoutActivity" />
         <activity android:name=".activities.MeActivity" />
         <activity android:name=".activities.EditProfileActivity" />
-        <activity android:name=".activities.SurveyActivity"></activity>
+        <activity android:name=".activities.SurveyActivity" />
         <activity
             android:name=".activities.MoreActivity"
-            android:hardwareAccelerated="false">
-        </activity>
+            android:hardwareAccelerated="false"></activity>
+        <activity android:name=".activities.QuickStatsActivity"></activity>
     </application>
 
 </manifest>
