+++ app/src/main/java/com/excy/excy/activities/QuickStatsActivity.java
@@ -1,0 +1,80 @@
+package com.excy.excy.activities;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+
+import com.excy.excy.R;
+import com.excy.excy.models.RVAdapter;
+import com.excy.excy.models.Workout;
+import com.excy.excy.utilities.AppUtilities;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+import java.util.ArrayList;
+
+import static com.excy.excy.activities.PlayActivity.getContext;
+
+public class QuickStatsActivity extends AppCompatActivity {
+    private ArrayList<Workout> workoutList;
+    private int workoutListSize = 20;
+    private int count = 0;
+    RVAdapter mAdapter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_quick_stats);
+
+        if (workoutList == null) {
+            workoutList = new ArrayList<>(workoutListSize);
+        }
+
+        RecyclerView mRecyclerView = (RecyclerView) findViewById(R.id.rvRecentWorkouts);
+        mRecyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        mAdapter = new RVAdapter(workoutList);
+        mRecyclerView.setAdapter(mAdapter);
+        mRecyclerView.setHasFixedSize(true);
+
+        String userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
+        DatabaseReference mDatabaseRef = FirebaseDatabase.getInstance()
+                .getReference(AppUtilities.TABLE_NAME_WORKOUTS + "/" + userId);
+
+        mDatabaseRef.addChildEventListener(new ChildEventListener() {
+            @Override
+            public void onChildAdded(DataSnapshot dataSnapshot, String prevChildKey) {
+                Workout workout = dataSnapshot.getValue(Workout.class);
+                workout.setId(dataSnapshot.getKey());
+                workoutList.add(workout);
+                count++;
+
+                if (count > workoutListSize) {
+                    workoutList.remove(0);
+                }
+
+                mAdapter.notifyDataSetChanged();
+            }
+
+            @Override
+            public void onChildChanged(DataSnapshot dataSnapshot, String prevChildKey) {}
+
+            @Override
+            public void onChildRemoved(DataSnapshot dataSnapshot) {
+                Workout workout = dataSnapshot.getValue(Workout.class);
+                workoutList.remove(workout);
+                mAdapter.notifyDataSetChanged();
+            }
+
+            @Override
+            public void onChildMoved(DataSnapshot dataSnapshot, String prevChildKey) {}
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {}
+        });
+    }
+}
