+++ app/src/main/res/layout/activity_me.xml
@@ -209,8 +209,9 @@
         <TextView
             android:id="@+id/tvRecentHistory"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="40dp"
             android:layout_below="@id/llMyInspiration"
+            android:gravity="center"
             android:layout_marginBottom="8dp"
             android:textAlignment="center"
             android:text="@string/recent_history"
